+++ lib/utilis/app_colors.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+class AppColor {
+  static Color kPrimary = const Color(0XFF1460F2);
+  static Color kWhite = const Color(0XFFFFFFFF);
+  static Color kBackground = const Color(0XFFFAFAFA);
+  static Color kGrayscaleDark100 = const Color(0XFF1C1C1E);
+  static Color kLine = const Color(0XFFEBEBEB);
+  static Color kBackground2 = const Color(0XFFF6F6F6);
+  static Color kGrayscale40 = const Color(0XFFAEAEB2);
+}
