+++ lib/utilis/image_assets.dart
@@ -1,0 +1,3 @@
+class ImagesPath {
+  static String kGoogleIcon = 'assets/images/googleSymbol.png';
+}
